requirements.txt
python-docx>=1.1.0
colorama>=0.4.6